# cakeshop-api/src/main/java/com/jpmorgan/cakeshop/model/Account.java
package com.jpmorgan.cakeshop.model;

import java.io.Serializable;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

import org.apache.commons.lang3.builder.ToStringBuilder;
import org.apache.commons.lang3.builder.ToStringStyle;

@Entity
@Table(name="ACCOUNTS")
public class Account implements Serializable {

    /**
     *
     */
    private static final long serialVersionUID = 4598804614905113573L;

    /**
     * Ethereum account address
     */
    @Id
    String address;

    String balance;

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getBalance() {
        return balance;
    }

    public void setBalance(String balance) {
        this.balance = balance;
    }

    @Override
    public String toString() {
        return ToStringBuilder.reflectionToString(this, ToStringStyle.MULTI_LINE_STYLE);
    }
}