# cakeshop-api/src/main/webapp/api/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "Cakeshop REST API",
		"version": "1.0.0",
		"contact": {}
	},
	"host": "localhost:8080",
	"schemes": [
		"http"
	],
	"basePath": "/cakeshop/api",
...
	"paths": {
...
		"/node/get": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Get node information",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "Node ID"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'node'"
										},
										"attributes": {
											"$ref": "#/definitions/Node"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/node/peers": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Retrieve information on the peers connected to the node.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"type": {
											"type": "string",
											"description": "resource type identifier, 'peer'"
										},
										"attributes": {
											"$ref": "#/definitions/Peer"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/node/start": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Start node",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "boolean",
									"description": "Whether or not the node was started"
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/node/stop": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Stop node",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "boolean",
									"description": "Whether or not the node was stopped"
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/node/reset": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Reset the chain to block zero",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "boolean",
									"description": "Whether or not the chain was reset successfully"
								},
								"errors": {
									"description": "Only present if errors were found",
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/node/update": {
			"post": {
				"tags": [
					"node"
				],
				"description": "Update node configuration. May cause node to restart or even reset.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "logLevel",
						"type": "integer",
						"required": false,
						"description": "Set log level (0-6, 6 is most verbose)",
						"in": "body"
					},
					{
						"name": "networkId",
						"type": "integer",
						"required": false,
						"description": "Network ID (any integer value; node will restart)",
						"in": "body"
					},
					{
						"name": "identity",
						"type": "string",
						"required": false,
						"description": "Change node identity (node will restart)",
						"in": "body"
					},
					{
						"name": "committingTransactions",
						"type": "boolean",
						"required": false,
						"description": "Whether or not the node should commit incoming transactions (i.e., toggle mining)",
						"in": "body"
					},
					{
						"name": "extraParams",
						"type": "string",
						"required": false,
						"description": "Extra parameters to pass to the node (startup flags; node will restart)",
						"in": "body"
					},
					{
						"name": "genesisBlock",
						"type": "string",
						"required": false,
						"description": "Genesis block to use for the chain (modifying this will reset the chain)",
						"in": "body"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"attributes": {
											"$ref": "#/definitions/NodeInfo"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/wallet/create": {
			"post": {
				"tags": [
					"wallet"
				],
				"description": "Crete a new account on the node",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "Response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "Receipt ID"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'wallet'"
										},
										"attributes": {
											"$ref": "#/definitions/Account"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/wallet/list": {
			"post": {
				"tags": [
					"wallet"
				],
				"description": "List all accounts on the node",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "array",
									"description": "response data",
									"items": {
										"type": "object",
										"properties": {
											"type": {
												"type": "string",
												"description": "resource type identifier, 'wallet'"
											},
											"attributes": {
												"$ref": "#/definitions/Account"
											}
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		}
	},
	"definitions": {
		"Account": {
			"properties": {
				"address": {
					"type": "string",
					"description": "160-bit identifier"
				},
				"balance": {
					"type": "string",
					"description": "A scalar value equal to the number of Wei owned by this address."
				}
			}
		},
		"APIError": {
			"properties": {
				"id": {
					"type": "string",
					"description": "Error ID, if available"
				},
				"status": {
					"type": "string",
					"description": "HTTP status code"
				},
				"code": {
					"type": "string",
					"description": "API-specific error code, if available"
				},
				"title": {
					"type": "string",
					"description": "Short error description"
				},
				"detail": {
					"type": "string",
					"description": "Detailed error information"
				}
			}
		},
...
		"Node": {
			"properties": {
				"status": {
					"type": "string",
					"description": "Status of the node, it has two values \"running\" or \"stopped\""
				},
				"id": {
					"type": "string",
					"description": "Unique Node ID"
				},
				"nodeName": {
					"type": "string",
					"description": "Friendly node name (includes client and version info)"
				},
				"nodeUrl": {
					"type": "string",
					"description": "enode URI (Includes host IP and port number)"
				},
				"nodeIP": {
					"type": "string",
					"description": "IP address of the node"
				},
				"peerCount": {
					"type": "integer",
					"description": "Number of peers connected to the node"
				},
				"pendingTxn": {
					"type": "integer",
					"description": "Number of transaction in the queue and haven't been mined"
				},
				"mining": {
					"type": "boolean",
					"description": "Indicates weather the miner is running or not. It has a true or false values."
				},
				"latestBlock": {
					"type": "integer",
					"description": "A number indicating the most recent block that has been mined."
				}
			}
		},
		"Peer": {
			"properties": {
				"status": {
					"type": "string",
					"description": "Status of the node, it has two values \"running\" or \"stopped\""
				},
				"id": {
					"type": "string",
					"description": "Unique Node ID"
				},
				"nodeName": {
					"type": "string",
					"description": "Friendly node name (includes client and version info)"
				},
				"nodeUrl": {
					"type": "string",
					"description": "enode URI (Includes host IP and port number)"
				},
				"nodeIP": {
					"type": "string",
					"description": "IP address of the node"
				}
			}
		},
		"NodeInfo": {
			"properties": {
				"identity": {
					"type": "string",
					"description": "Friendly node name which gets included in the full node Name"
				},
				"committingTransactions": {
					"type": "boolean",
					"description": "Indicates whether the node is commiting transactions or not"
				},
				"networkId": {
					"type": "integer",
					"description": "Network identifier"
				},
				"logLevel": {
					"type": "integer",
					"description": "Logging verbosity: 0-6 (0=silent, 1=error, 2=warn, 3=info, 4=core, 5=debug, 6=debug detail)"
				}
			}
		}
	}
}